Fixed hard pinned dependencies in requirements and setup.py.
The requirements were pinned to a version and that was causing trouble
when used with other packages. For example, firefly was expecting
requests version 2.18.1 while the package where this is included was
expecting 2.19.1.

Fixed it by changing the constraint from `==` to `>=`.

Also made sure the requirements.txt and setup.py are in sync.